Create useNavigateTo hook and bundle a few of the accelerator routes
